+++ click_extra/tests/test_commands.py
@@ -200,11 +200,11 @@ def test_subcommand_help(invoke, all_command_cli, cmd_id, param):
 
     colored_help_header = (
         r"It works!\n"
-        rf"\x1b\[94m\x1b\[1mUsage: "
+        r"\x1b\[94m\x1b\[1m\x1b\[4mUsage: "
         rf"\x1b\[0m\x1b\[97mcommand-cli1 {cmd_id}-subcommand\x1b\[0m"
-        r" \x1b\[90m\[OPTIONS\]\x1b\[0m\n"
+        r" \x1b\[36m\x1b\[2m\[OPTIONS\]\x1b\[0m\n"
         r"\n"
-        r"\x1b\[94m\x1b\[1mOptions:\x1b\[0m\n"
+        r"\x1b\[94m\x1b\[1m\x1b\[4mOptions:\x1b\[0m\n"
     )
 
     # Extra sucommands are colored and include all extra options.