+++ changelog.md
@@ -6,6 +6,7 @@
 This version is not released yet and is under active development.
 ```
 
+- Color environment variables and defaults in help screens.
 - Allow setting the style of `(deprecated)` tag in help screen's theme. Defaults to bold, bright yellow.
 - Change default style of critical log messages' prefix to bold, bright red.